Median absolute deviation - Fix percentile calculation and refactor.
Fix the percentile calculation, which subtracted the mean of each value
before determining the value although it should have (and does now) take
the relative difference as is. Extended the method at the same time to
fetch quantile values for absolute and relative data values.
Replaced the internal caching hash with functools lru_cache.